+++ CHANGELOG.md
@@ -55,6 +55,8 @@ An EtherCAT master written in Rust.
 - [#169](https://github.com/ethercrab-rs/ethercrab/pull/169) Linux only: add `io_uring`-based
   blocking TX/RX loop for better performance.
 - [#173](https://github.com/ethercrab-rs/ethercrab/pull/173) Add MUSL libc support.
+- [#178](https://github.com/ethercrab-rs/ethercrab/pull/178) Add `Error::SubDevice` to get a
+  subdevice status code on failure.
 
 ### Fixed